backend/pyproject.toml
[tool.poetry]
name = "git-ranker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
phidata = "^2.5.33"
openai = "^1.55.3"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
fastapi = "0.115.4"
fastapi-cli = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
